+++ package.json
@@ -10,6 +10,7 @@
     "react-dom": "^17.0.2",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
+    "react-slick": "^0.28.1",
     "web-vitals": "^2.1.3"
   },
   "scripts": {
