+++ public/index.html
@@ -55,12 +55,12 @@
 
     <!-- End Top To Bottom Area -->
     <!-- JS ============================================ -->
-    <script defer src="assets/js/vendor/jquery.js"></script>
-    <script async src="assets/js/vendor/jquery.nice-select.min.js"></script>
+    <script  src="assets/js/vendor/jquery.js"></script>
+    <script defer src="assets/js/vendor/jquery.nice-select.min.js"></script>
     <script defer src="assets/js/vendor/jquery-ui.js"></script>
     <script defer src="assets/js/vendor/modernizer.min.js"></script>
     <script defer src="assets/js/vendor/feather.min.js"></script>
-    <script defer src="assets/js/vendor/slick.min.js"></script>
+    <!-- <script defer src="assets/js/vendor/slick.min.js"></script> -->
     <script defer src="assets/js/vendor/bootstrap.min.js"></script>
     <script defer src="assets/js/vendor/sal.min.js"></script>
     <script defer src="assets/js/vendor/waypoint.js"></script>
@@ -77,7 +77,7 @@
     <script defer src="assets/js/vendor/jquery.custom-file-input.js"></script>
 
     <!-- main JS -->
-    <script  src="assets/js/main.js"></script>
+    <script  defer src="assets/js/main.js"></script>
   </head>
   <body class="template-color-1">
     <noscript>You need to enable JavaScript to run this app.</noscript>
