+++ src/App.js
@@ -1,7 +1,7 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import Layout from "./components/common/Layout";
 import About from "./pages/About";
-import Home from "./pages/Home";
+import Home from "./pages/home/Home";
 
 function App() {
   return (
